Refactor Checking to Rules
