+++ test/test_files/test_error_expected/02_library/test_dummy.vhd
@@ -1,7 +1,7 @@
 -- Dummy for other test cases.
 entity test is
   port (
-    i : in std_logic -- dummy error
+    i_a : in std_ulogic -- dummy error
   );
 end entity;
 context test_ContextB is
