+++ test/test_files/test_error_expected/07_ignoring/ignore_no_read.vhd
@@ -1,0 +1,10 @@
+entity test is
+end test;
+architecture arch of test is
+  -- Warning should be unused, ignoring should not ignore rule
+  --vhdl-linter-disable-next-line not-declared
+  signal test : integer;
+
+begin
+test <= 0;
+end arch ;
