+++ test/test_files/test_error_expected/07_ignoring/ignore_no_read_2.vhd
@@ -1,0 +1,9 @@
+entity test is
+end test;
+architecture arch of test is
+  -- Warning should be unused, ignoring should not ignore rule
+  signal test : integer; --vhdl-linter-disable-line not-declared
+
+begin
+test <= 0;
+end arch ;
