+++ test/test_files/test_error_expected/07_ignoring/ignore_no_read_4.vhd
@@ -1,0 +1,12 @@
+--vhdl-linter-disable unused
+entity test is
+end test;
+-- unspecificed enabling should enable all rules
+--vhdl-linter-enable
+architecture arch of test is
+  -- Warning activated again, ignoring should not ignore rule
+  signal test : integer;
+
+begin
+  test <= 0;
+end arch;
