+++ test/test_files/test_no_error/07_ignoring/ignore_no_read.vhd
@@ -1,0 +1,9 @@
+entity test is
+end test;
+architecture arch of test is
+  --vhdl-linter-disable-next-line unused
+  signal test : integer;
+
+begin
+test <= 0;
+end arch ;
