+++ test/test_files/test_no_error/07_ignoring/ignore_no_read_3.vhd
@@ -1,0 +1,9 @@
+--vhdl-linter-disable unused
+entity test is
+end test;
+architecture arch of test is
+  signal test : integer;
+
+begin
+  test <= 0;
+end arch;
