+++ test/test_files/test_no_error/07_ignoring/ignore_no_read_4.vhd
@@ -1,0 +1,9 @@
+--vhdl-linter-disable
+entity test is
+end test;
+architecture arch of test is
+  signal test : integer;
+
+begin
+  test <= 0;
+end arch;
